requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
gtfs-realtime-bindings==1.0.0
idna==3.10
protobuf==6.32.1
psycopg2-binary==2.9.10
requests==2.32.5
setuptools==80.9.0
urllib3==2.5.0
